feat: 혜택 미리보기 Controller 추가
- 기존 Reservation 기능을 PreviewBenefit으로 이동
