+++ src/main/java/christmas/controller/PreviewBenefit.java
@@ -1,0 +1,39 @@
+package christmas.controller;
+
+public class PreviewBenefit {
+    public PreviewBenefit(){
+        order();
+        benefit();
+        discountPayment();
+        badge();
+    }
+
+    public void order() {
+        menu();
+        payment();
+    }
+
+    public void menu() {
+
+    }
+
+    public void payment() {
+    }
+
+    public void benefit() {
+        presentation();
+        totalBenefit();
+    }
+
+    public void presentation() {
+    }
+
+    public void totalBenefit() {
+    }
+
+    public void discountPayment() {
+    }
+
+    public void badge() {
+    }
+}
